Remove redundant judgments
I deleted the code `event.target! == el`. The reason is as follows:

- For `path.indexOf(el)`

If `el` is not in `event.path`，indicates that the click occurred outside `el`

- For `!el.contains(event.target)`

If `event.target` is not a descendant element of `el` or `el` itself，indicates that the click occurred outside `el`

Therefore, the additional judgment of `event.target! == el` is not needed.

(Forgive my English is not good enough 😂)